+++ src/cpp/rtps/writer/BaseWriter.hpp
@@ -419,6 +419,30 @@ protected:
 
     //! Liveliness lost status of this writer
     LivelinessLostStatus liveliness_lost_status_;
+    /// Is the data sent directly or announced by HB and THEN sent to the ones who ask for it?.
+    bool push_mode_ = true;
+
+    /// Flow controller.
+    FlowController* flow_controller_;
+    /// Maximum number of bytes allowed for an RTPS datagram generated by this writer.
+    uint32_t max_output_message_size_ = std::numeric_limits<uint32_t>::max();
+
+    /// WriterHistory
+    WriterHistory* history_ = nullptr;
+    /// Listener
+    WriterListener* listener_ = nullptr;
+    /// Asynchronous publication activated
+    bool is_async_ = false;
+    /// Separate sending activated
+    bool separate_sending_enabled_ = false;
+
+    /// The liveliness kind of this writer
+    dds::LivelinessQosPolicyKind liveliness_kind_;
+    /// The liveliness lease duration of this writer
+    Duration_t liveliness_lease_duration_;
+    /// The liveliness announcement period
+    Duration_t liveliness_announcement_period_;
+
 
 };
