+++ test/mock/rtps/RTPSWriter/rtps/writer/BaseWriter.hpp
@@ -42,6 +42,8 @@ public:
     {
     }
 
+    virtual ~BaseWriter() = default;
+
     static BaseWriter* downcast(
             RTPSWriter* writer)
     {
